chore: rename migrator to mtma; provide notes on intentions.
