+++ checks/migrator/checks/sketchpad/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "migration-migrator-checks-sketchpad"
+name = "mtma-migrator-checks-sketchpad"
 version = { workspace = true }
 edition = { workspace = true }
 license = { workspace = true }
@@ -9,16 +9,16 @@ publish = { workspace = true }
 rust-version = { workspace = true }
 
 [dependencies]
-migration-migrator-types = { workspace = true }
-migration-migrator-test-types = { workspace = true }
+mtma-migrator-types = { workspace = true }
+mtma-migrator-test-types = { workspace = true }
 anyhow = { workspace = true }
 mtma-node-null-core = { workspace = true }
 tokio = { workspace = true }
 chrono = { workspace = true }
 rand = { workspace = true }
 kestrel = { workspace = true }
-migration-node-preludes = { workspace = true }
-migration-node-test-types = { workspace = true }
+mtma-node-preludes = { workspace = true }
+mtma-node-test-types = { workspace = true }
 
 [dev-dependencies]
 tracing-test = { workspace = true }
