+++ checks/node/checks/movement-mainnet/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "migration-node-checks-mainnet"
+name = "mtma-node-checks-mainnet"
 version = { workspace = true }
 edition = { workspace = true }
 license = { workspace = true }
@@ -9,13 +9,13 @@ publish = { workspace = true }
 rust-version = { workspace = true }
 
 [dependencies]
-migration-node-test-types = { workspace = true }
+mtma-node-test-types = { workspace = true }
 anyhow = { workspace = true }
 mtma-node-null-core = { workspace = true }
-migration-node-test-global-storage-injective-criterion = { workspace = true }
-migration-node-test-global-storage-includes-criterion = { workspace = true }
-migration-node-test-global-storage-not-empty-criterion = { workspace = true }
-migration-node-preludes = { workspace = true }
+mtma-node-test-global-storage-injective-criterion = { workspace = true }
+mtma-node-test-global-storage-includes-criterion = { workspace = true }
+mtma-node-test-global-storage-not-empty-criterion = { workspace = true }
+mtma-node-preludes = { workspace = true }
 movement-syncing = { workspace = true }
 maptos-execution-util = { workspace = true }
 tokio = { workspace = true }
