+++ checks/node/checks/movement-mainnet/src/global_storage_includes.rs
@@ -3,8 +3,8 @@ pub mod test {
 
 	use futures::channel::mpsc as futures_mpsc;
 	use maptos_execution_util::config::Config as MaptosConfig;
-	use migration_node_test_global_storage_includes_criterion::GlobalStorageIncludes;
-	use migration_node_test_types::{
+	use mtma_node_test_global_storage_includes_criterion::GlobalStorageIncludes;
+	use mtma_node_test_types::{
 		check::checked_migration,
 		criterion::movement_executor::{MovementNode, MovementOptExecutor},
 		prelude::Prelude,
