+++ checks/node/checks/sketchpad/src/global_storage_not_empty.rs
@@ -1,9 +1,9 @@
 #[cfg(test)]
 pub mod test {
 
-	use migration_node_preludes::basic::BasicPrelude;
-	use migration_node_test_global_storage_not_empty_criterion::GlobalStorageNotEmpty;
-	use migration_node_test_types::{
+	use mtma_node_preludes::basic::BasicPrelude;
+	use mtma_node_test_global_storage_not_empty_criterion::GlobalStorageNotEmpty;
+	use mtma_node_test_types::{
 		check::checked_migration,
 		criterion::movement_executor::{MovementNode, MovementOptExecutor},
 		prelude::PreludeGenerator,
