+++ checks/node/citeria/global-storage-injective/src/lib.rs
@@ -1,9 +1,9 @@
-use migration_node_test_types::criterion::movement_aptos_executor::TStateView as _;
-use migration_node_test_types::criterion::movement_executor::TStateView as _;
-use migration_node_test_types::criterion::{
+use mtma_node_test_types::criterion::movement_aptos_executor::TStateView as _;
+use mtma_node_test_types::criterion::movement_executor::TStateView as _;
+use mtma_node_test_types::criterion::{
 	Criterion, CriterionError, Criterionish, MovementAptosNode, MovementNode,
 };
-use migration_node_test_types::criterion::movement_aptos_executor::aptos_types::state_store::state_key::StateKey as MovementAptosStateKey;
+use mtma_node_test_types::criterion::movement_aptos_executor::aptos_types::state_store::state_key::StateKey as MovementAptosStateKey;
 use tracing::debug;
 pub struct GlobalStorageInjective;
