+++ checks/node/util/types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "migration-node-test-types"
+name = "mtma-node-test-types"
 version = { workspace = true }
 edition = { workspace = true }
 license = { workspace = true }
@@ -19,7 +19,7 @@ aptos-types = { workspace = true }
 thiserror = { workspace = true }
 aptos-storage-interface = { workspace = true }
 either = { workspace = true }
-migration-node-types = { workspace = true }
+mtma-node-types = { workspace = true }
 chrono = { workspace = true }
 
 [lints]
