+++ migration/core/migrator/mtma-null/Cargo.toml
@@ -17,7 +17,7 @@ anyhow = { workspace = true }
 clap-markdown-ext = { workspace = true }
 thiserror = { workspace = true }
 orfile = { workspace = true }
-migration-node-types = { workspace = true }
+mtma-node-types = { workspace = true }
 aptos-db = { workspace = true }
 aptos-storage-interface = { workspace = true }
 aptos-config = { workspace = true }
