+++ migration/util/migrator-types/src/migrator/movement_migrator.rs
@@ -2,7 +2,7 @@ use anyhow::Context;
 use kestrel::WaitCondition;
 pub use maptos_opt_executor;
 pub use maptos_opt_executor::aptos_types::{chain_id::ChainId, state_store::TStateView};
-use migration_node_types::executor::MovementNode;
+use mtma_node_types::executor::MovementNode;
 use movement_client::rest_client::Client as MovementRestClient;
 use movement_core::Movement;
