+++ migration/util/node-types/README.md
@@ -1,0 +1,2 @@
+# `migration-node-types`
+Types defining and for working with the a complete "node."
